RDKB-16458: 2.9p3s1 Not upgrading to latest Firmware 2.9p5s1
Reason for change: Xconf scripts were exiting due to an error in maintenance window compute.
LTime returns Hour/Minute with a "0" appended which shell interpreted as octal value when returned values are 08 or 09.
Fix is to remove leading zero from the returned value of LTime.

Test Procedure: Modified system time so that LTime will return 08 or 09 as hour/minute
Ran the maintenance window computation logic and verified that the time is calculated properly.
Risks: Low
Signed-off-by: Anoop Nair Chelakkode<anoopnair_c@infosys.com>

Change-Id: I3bc09d2e9a7352787020aca442bdeb1523afe09d